// src/app/core/models/reading-models/reading-filter.model.ts
import { PaginationFilter } from "../shared-models/pagination.model";

export class ReadingFilter extends PaginationFilter {
    readingMainLevelId: number = 0;
    readingSubLevelId: number = 0;
    learningOutcomeId: number = 0;
    gradeId: number = 0;
    subjectId: number = 0;
    learningHubId: number = 0;
    ibLearnerProfileId: number = 0;
    genreId: number = 0;
    authorId: number = 0;
    publisherId: number = 0;
    illustratorId: number = 0;
    ageGroupId: number = 0;
  }
  